Add a facility to set local/remote video output mode
The current DRM stack requres us to explicitly set the video output
mode between local (BMC) and remote (host). We need to set this mode on
render start/stop, so that the output is set correctly.

This change adds another argument to specify a sysfs file, into which we
set the output mode.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>