Fix error when VGA scratch register has high bits set
Currently, we expect the VGA scratch register to only have bits set
in the least-significant byte (the password byte), and use the
presence of other bits to indicate a strtoul() failure.

However, having high bits set isn't necessarily an error, as this is a
32-bit register, and the other scratch register values may be set
arbitrarily.

This change makes the strtoul() checking more tolerant of scratch
register values.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>